+++ topk-cli/Cargo.toml
@@ -71,8 +71,7 @@ colored = "2"
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "json"] }
 sha2 = "0.10"
 base64 = "0.22"
-rand = "0.9"
-keyring = "4"
+oauth2 = { version = "5", default-features = false, features = ["reqwest", "timing-resistant-secret-traits"] }
 tempfile = "3"
 
 [dev-dependencies]
