+++ topk-cli/src/auth/callback.rs
@@ -9,6 +9,7 @@ use axum::routing::get;
 use axum::{serve, Router};
 use base64::engine::general_purpose::STANDARD;
 use base64::Engine;
+use oauth2::CsrfToken;
 use serde::Deserialize;
 use tokio::net::TcpListener;
 use tokio::sync::{mpsc, oneshot};
@@ -20,7 +21,7 @@ const SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(10);
 /// The server stays scoped to this future so cancellation releases the listener.
 pub(super) async fn run<F, Fut, T>(
     listener: TcpListener,
-    state: String,
+    state: CsrfToken,
     duration: Duration,
     complete: F,
 ) -> Result<T>
@@ -64,7 +65,7 @@ where
     result
 }
 
-fn router(state: String) -> (Router, mpsc::Receiver<Callback>) {
+fn router(state: CsrfToken) -> (Router, mpsc::Receiver<Callback>) {
     let (callbacks, receiver) = mpsc::channel(1);
     let router = Router::new()
         .route(
@@ -85,7 +86,7 @@ struct CallbackParams {
 
 #[derive(Clone)]
 struct CallbackState {
-    state: String,
+    state: CsrfToken,
     callbacks: mpsc::Sender<Callback>,
 }
 
@@ -99,7 +100,7 @@ async fn callback(
     State(state): State<CallbackState>,
     Query(params): Query<CallbackParams>,
 ) -> Response {
-    if params.state != state.state {
+    if CsrfToken::new(params.state) != state.state {
         return StatusCode::BAD_REQUEST.into_response();
     }
     let code = match (params.code, params.error) {
@@ -154,5 +155,36 @@ fn page(ok: bool) -> Response {
 }
 
 #[cfg(test)]
-#[path = "../../tests/auth/callback.rs"]
-mod tests;
+mod tests {
+    use std::time::Duration;
+
+    use oauth2::CsrfToken;
+    use tokio::net::{TcpListener, TcpStream};
+
+    use super::run;
+
+    #[tokio::test]
+    async fn timeout_releases_listener_even_with_a_silent_connection() {
+        for connected in [false, true] {
+            let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
+            let addr = listener.local_addr().unwrap();
+            let _connection = if connected {
+                Some(TcpStream::connect(addr).await.unwrap())
+            } else {
+                None
+            };
+            let error = run::<_, _, ()>(
+                listener,
+                CsrfToken::new("test-state".into()),
+                Duration::from_millis(50),
+                |_| async {
+                    unreachable!("unexpected callback");
+                },
+            )
+            .await
+            .unwrap_err();
+            assert!(error.to_string().contains("timed out"));
+            assert!(TcpListener::bind(addr).await.is_ok());
+        }
+    }
+}
