+++ topk-cli/src/auth/login.rs
@@ -1,29 +1,30 @@
+use std::borrow::Cow;
 use std::net::SocketAddr;
 use std::time::Duration;
 
 use anyhow::{Context, Result};
+use oauth2::{
+    AuthorizationCode, CsrfToken, PkceCodeChallenge, PkceCodeVerifier, RedirectUrl, Scope,
+};
 use tokio::net::TcpListener;
 use url::Url;
 
 use super::callback;
 use super::session::Session;
-use super::util::{random_urlsafe, s256};
 use super::{AccessTokenClaims, Auth};
 
 // Configured callback URLs (127.0.0.1) ports
 const PORTS: [u16; 3] = [38123, 38124, 38125];
-// OAuth scope for the login
-const SCOPE: &str = "openid profile email offline_access";
-// Timeout for the login
+const SCOPES: [&str; 4] = ["openid", "profile", "email", "offline_access"];
 const TIMEOUT: Duration = Duration::from_secs(5 * 60);
 
 pub struct Login<'a> {
     auth: &'a Auth,
     listener: TcpListener,
     url: Url,
-    redirect_uri: String,
-    verifier: String,
-    state: String,
+    redirect_uri: RedirectUrl,
+    verifier: PkceCodeVerifier,
+    state: CsrfToken,
 }
 
 impl<'a> Login<'a> {
@@ -32,24 +33,20 @@ impl<'a> Login<'a> {
         let listener = TcpListener::bind(addresses.as_slice())
             .await
             .with_context(|| format!("cannot listen on 127.0.0.1 ports {PORTS:?}"))?;
-        let verifier = random_urlsafe();
-        let state = random_urlsafe();
-        let redirect_uri = format!(
+        let (challenge, verifier) = PkceCodeChallenge::new_random_sha256();
+        let redirect_uri = RedirectUrl::new(format!(
             "http://127.0.0.1:{}/callback",
             listener.local_addr()?.port()
-        );
-        let mut url = auth.oauth_config.issuer.join("authorize")?;
-        url.query_pairs_mut().extend_pairs([
-            ("response_type", "code"),
-            ("client_id", auth.oauth_config.client_id.as_str()),
-            ("redirect_uri", &redirect_uri),
-            ("scope", SCOPE),
-            ("audience", &auth.oauth_config.audience),
-            ("code_challenge", &s256(&verifier)),
-            ("code_challenge_method", "S256"),
-            ("state", &state),
-            ("prompt", "login"),
-        ]);
+        ))?;
+        let (url, state) = auth
+            .client
+            .authorize_url(CsrfToken::new_random)
+            .set_redirect_uri(Cow::Borrowed(&redirect_uri))
+            .add_scopes(SCOPES.map(|scope| Scope::new(scope.to_string())))
+            .add_extra_param("audience", &auth.oauth_config.audience)
+            .add_extra_param("prompt", "login")
+            .set_pkce_challenge(challenge)
+            .url();
         Ok(Self {
             auth,
             listener,
@@ -69,24 +66,18 @@ impl<'a> Login<'a> {
             let res = self
                 .auth
                 .client
-                .post_token(&[
-                    ("grant_type", "authorization_code"),
-                    ("client_id", &self.auth.oauth_config.client_id),
-                    ("code", &code),
-                    ("code_verifier", &self.verifier),
-                    ("redirect_uri", &self.redirect_uri),
-                ])
+                .exchange_code(AuthorizationCode::new(code))
+                .set_pkce_verifier(self.verifier)
+                .set_redirect_uri(Cow::Owned(self.redirect_uri))
+                .request_async(&self.auth.http)
                 .await
                 .context("exchanging the authorization code")?;
-            let session = Session::from_response(res, None);
-            let claims = AccessTokenClaims::parse(&session.access_token, self.auth.audience());
+            let session = Session::from_response(res, None)?;
+            let claims =
+                AccessTokenClaims::parse(&session.access_token, &self.auth.oauth_config.audience);
             self.auth.store.lock().await?.save(session)?;
             Ok(claims)
         })
         .await
     }
 }
-
-#[cfg(test)]
-#[path = "../../tests/auth/login.rs"]
-mod tests;
