--- topk-cli/src/auth/util.rs
@@ -1,18 +1,0 @@
-use base64::engine::general_purpose::URL_SAFE_NO_PAD;
-use base64::Engine;
-use rand::RngCore;
-use sha2::{Digest, Sha256};
-
-pub(super) fn s256(verifier: &str) -> String {
-    URL_SAFE_NO_PAD.encode(Sha256::digest(verifier.as_bytes()))
-}
-
-pub(super) fn random_urlsafe() -> String {
-    let mut bytes = [0u8; 32];
-    rand::rng().fill_bytes(&mut bytes);
-    URL_SAFE_NO_PAD.encode(bytes)
-}
-
-#[cfg(test)]
-#[path = "../../tests/auth/util.rs"]
-mod tests;
