+++ topk-cli/src/commands/mod.rs
@@ -1,2 +1,3 @@
 #[cfg(feature = "import")]
 pub mod import;
+pub mod login;
