+++ topk-cli/src/lib.rs
@@ -2,5 +2,6 @@ pub mod auth;
 pub mod commands;
 pub mod config;
 pub mod endpoint;
+pub mod file;
 #[cfg(feature = "import")]
 pub mod import;
