+++ topk-cli/tests/auth/login.rs
@@ -1,10 +1,14 @@
+use std::collections::HashMap;
+use std::time::Duration;
+
+use oauth2::{PkceCodeChallenge, PkceCodeVerifier};
 use tempfile::TempDir;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::TcpStream;
+use tokio::net::{TcpListener, TcpStream};
 use tokio::time::{sleep, timeout};
+use url::Url;
 
-use super::*;
-use crate::auth::tests::{response, Server, LOGIN};
+use super::common::{response, Server, LOGIN};
 
 async fn request(port: u16, method: &str, path: &str) -> String {
     let mut stream = TcpStream::connect(("127.0.0.1", port)).await.unwrap();
@@ -29,11 +33,19 @@ async fn unrelated_requests_do_not_finish_login_and_success_follows_persistence(
     let mut server = Server::new().await;
     let auth = server.auth(dir.path());
     let _guard = LOGIN.lock().await;
-    let login = Login::new(&auth).await.unwrap();
-    let port = login.listener.local_addr().unwrap().port();
-    let path = format!("/callback?code=valid&state={}", login.state);
-    let empty = format!("/callback?code=&state={}", login.state);
-    let duplicate = format!("{path}&state={}", login.state);
+    let login = auth.login().await.unwrap();
+    let params: HashMap<_, _> = login.url().query_pairs().into_owned().collect();
+    let callback = Url::parse(&params["redirect_uri"]).unwrap();
+    assert_eq!(params["response_type"], "code");
+    assert_eq!(params["client_id"], server.config().client_id);
+    assert_eq!(params["audience"], server.config().audience);
+    assert_eq!(params["scope"], "openid profile email offline_access");
+    assert_eq!(params["prompt"], "login");
+    assert_eq!(params["code_challenge_method"], "S256");
+    let port = callback.port().unwrap();
+    let path = format!("/callback?code=valid&state={}", params["state"]);
+    let empty = format!("/callback?code=&state={}", params["state"]);
+    let duplicate = format!("{path}&state={}", params["state"]);
     let ambiguous = format!("{path}&error=access_denied");
     let mut silent = TcpStream::connect(("127.0.0.1", port)).await.unwrap();
     let browser = async {
@@ -55,8 +67,21 @@ async fn unrelated_requests_do_not_finish_login_and_success_follows_persistence(
         let callback = request(port, "GET", &path);
         let exchange = async {
             let req = server.request().await;
-            assert!(req.contains("code=valid") && req.contains("code_verifier="));
-            assert!(auth.store.lock().await.unwrap().load().unwrap().is_none());
+            let form: HashMap<_, _> = url::form_urlencoded::parse(req.as_bytes())
+                .into_owned()
+                .collect();
+            assert_eq!(form["grant_type"], "authorization_code");
+            assert_eq!(form["code"], "valid");
+            assert_eq!(form["client_id"], params["client_id"]);
+            assert_eq!(form["redirect_uri"], params["redirect_uri"]);
+            assert_eq!(
+                PkceCodeChallenge::from_code_verifier_sha256(&PkceCodeVerifier::new(
+                    form["code_verifier"].clone(),
+                ))
+                .as_str(),
+                params["code_challenge"],
+            );
+            assert!(!dir.path().join("credentials.toml").exists());
             assert!(request(port, "GET", &path)
                 .await
                 .starts_with("HTTP/1.1 409"));
@@ -84,9 +109,11 @@ async fn failed_exchange_and_failed_save_never_send_success() {
         let mut server = Server::new().await;
         let auth = server.auth(dir.path());
         let _guard = LOGIN.lock().await;
-        let login = Login::new(&auth).await.unwrap();
-        let port = login.listener.local_addr().unwrap().port();
-        let path = format!("/callback?code=valid&state={}", login.state);
+        let login = auth.login().await.unwrap();
+        let params: HashMap<_, _> = login.url().query_pairs().into_owned().collect();
+        let callback = Url::parse(&params["redirect_uri"]).unwrap();
+        let port = callback.port().unwrap();
+        let path = format!("/callback?code=valid&state={}", params["state"]);
         if fail_save {
             // A real filesystem failure, without changing the storage implementation.
             std::fs::create_dir(dir.path().join("credentials.toml")).unwrap();
@@ -126,8 +153,10 @@ async fn cancelling_login_drops_listener() {
     let server = Server::new().await;
     let auth = server.auth(dir.path());
     let _guard = LOGIN.lock().await;
-    let login = Login::new(&auth).await.unwrap();
-    let addr = login.listener.local_addr().unwrap();
+    let login = auth.login().await.unwrap();
+    let params: HashMap<_, _> = login.url().query_pairs().into_owned().collect();
+    let callback = Url::parse(&params["redirect_uri"]).unwrap();
+    let addr = ("127.0.0.1", callback.port().unwrap());
     tokio::select! {
         _ = login.finish() => panic!("unexpected callback"),
         _ = sleep(Duration::from_millis(25)) => {},
@@ -141,11 +170,13 @@ async fn authenticated_denial_finishes_without_exchanging_a_token() {
     let mut server = Server::new().await;
     let auth = server.auth(dir.path());
     let _guard = LOGIN.lock().await;
-    let login = Login::new(&auth).await.unwrap();
-    let port = login.listener.local_addr().unwrap().port();
+    let login = auth.login().await.unwrap();
+    let params: HashMap<_, _> = login.url().query_pairs().into_owned().collect();
+    let callback = Url::parse(&params["redirect_uri"]).unwrap();
+    let port = callback.port().unwrap();
     let path = format!(
         "/callback?state={}&error=access_denied&error_description=cancelled",
-        login.state
+        params["state"]
     );
     let (result, page) = tokio::join!(login.finish(), request(port, "GET", &path));
     assert!(result
