+++ topk-cli/tests/auth/main.rs
@@ -1,3 +1,4 @@
-// CLI integration tests. Private-module tests alongside these are included by
-// their owning modules in src/auth and run as library unit tests.
 mod cli;
+mod common;
+mod login;
+mod session;
